+++ src/main/java/com/noraknorak/user/service/UserRegisterService.java
@@ -78,30 +78,4 @@ public class UserRegisterService {
         }
         throw UserErrorCode.NOT_EQUAL_CODE.toException();
     }
-
-    // 부모/자식 연동 코드 검증
-    public User validateUserCode(String code) {
-        return userRepository.findByUserCode(code)
-                .orElseThrow(UserErrorCode.NOT_EQUAL_USER_CODE::toException);
-    }
-
-    // 부모/자식 관계 설정
-    @Transactional
-    public void verifyRelatedUser(CustomUserDetails customUserDetails, String role, Long relatedUserId){
-        Long userId = customUserDetails.getId();
-        try{
-            if(Role.SENIOR.getName().equals(role)){
-                userRepository.updateUserRoleAndRelatedUser(userId, Role.GUARDIAN, relatedUserId);
-            }else if(Role.GUARDIAN.getName().equals(role)){
-                userRepository.updateUserRoleAndRelatedUser(userId, Role.SENIOR, relatedUserId);
-            }else{
-                throw UserErrorCode.INVALID_ROLE_ERROR.toException();
-            }
-        } catch (DomainException e) {
-            throw e;
-        }
-        catch (Exception e){
-            throw UserErrorCode.INTERNAL_SERVER_ERROR.toException();
-        }
-    }
 }
